// src/gql/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Estimate = Timestamps & {
  __typename?: 'Estimate';
  createdAt: Scalars['ISO8601DateTime']['output'];
  updatedAt: Scalars['ISO8601DateTime']['output'];
  value: Scalars['Float']['output'];
};
// ...
export type Issue = ActivityFeedField & GithubTimestamps & Node & Timestamps & ViewerPermission & {
  __typename?: 'Issue';
  activityFeed?: Maybe<ActivityFeedConnection>;
  assignees: UserConnection;
  /** Issues that are blocked by this Issue. */
  blockedIssues: IssueConnection;
  /** Dependent items that are blocked by this issue */
  blockedItems: IssueDependencyItemConnection;
  /** Issues that are blocking this Issue. */
  blockingIssues: IssueConnection;
  /** Dependent items that are blocking this issue */
  blockingItems: IssueDependencyItemConnection;
  body?: Maybe<Scalars['String']['output']>;
  closedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  comments: CommentConnection;
  /** Return the Pull Requests connected to this issue */
  connectedPrs: IssueConnection;
  /** 1st step connections of the PR. Only available for PRs */
  connections: IssueConnection;
  createdAt: Scalars['ISO8601DateTime']['output'];
  /** @deprecated This field must not be used. */
  epic?: Maybe<Epic>;
  estimate?: Maybe<Estimate>;
  /** Estimates given by participants */
  estimationVotes: EstimationVoteConnection;
  ghCreatedAt: Scalars['ISO8601DateTime']['output'];
  ghId?: Maybe<Scalars['Int']['output']>;
  ghNodeId?: Maybe<Scalars['ID']['output']>;
  ghUpdatedAt: Scalars['ISO8601DateTime']['output'];
  htmlBody?: Maybe<Scalars['String']['output']>;
  htmlUrl: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  labels: LabelConnection;
  milestone?: Maybe<Milestone>;
  number: Scalars['Int']['output'];
  /** @deprecated This field must not be used. Use parentZenhubEpics. */
  parentEpics: EpicConnection;
  parentZenhubEpics: ZenhubEpicConnection;
  /** Issue connections to different pipelines */
  pipelineIssue?: Maybe<PipelineIssue>;
  pipelineIssues: PipelineIssueConnection;
  pullRequest: Scalars['Boolean']['output'];
  pullRequestObject?: Maybe<PullRequest>;
  pullRequestReviews?: Maybe<PullRequestReviewConnection>;
  releases: ReleaseConnection;
  repository: Repository;
  reviewRequests?: Maybe<ReviewRequestConnection>;
  /** Sprints that the issue belongs to */
  sprints: SprintConnection;
  state: IssueState;
  timelineItems: TimelineItemConnection;
  title: Scalars['String']['output'];
  /** Github or Zebhub issue */
  type: IssueType;
  updatedAt: Scalars['ISO8601DateTime']['output'];
  /** Issue creator; Only available for GitHub Issues */
  user: Owner;
  viewerPermission: PermissionLevel;
  zenhubAssignees: ZenhubUserConnection;
  zenhubLabels: ZenhubLabelConnection;
};
// ...
export type IssueConnection = {
  __typename?: 'IssueConnection';
  /** A list of edges. */
  edges: Array<IssueEdge>;
  /** A list of nodes. */
  nodes: Array<Issue>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  pipelineCounts: PipelineCounts;
  /** @deprecated Use common field pipelineCounts instead */
  splitCounts: Scalars['JSON']['output'];
  /** @deprecated Use common field pipelineCounts instead */
  sumEstimates: Scalars['Float']['output'];
  totalCount: Scalars['Int']['output'];
};
// ...
export type IssueEdge = {
  __typename?: 'IssueEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['String']['output'];
  /** The item at the end of the edge. */
  node: Issue;
};
// ...
export enum IssueState {
  Closed = 'CLOSED',
  Open = 'OPEN'
}
// ...
export type Milestone = GithubTimestamps & Timestamps & {
  __typename?: 'Milestone';
  closedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  closedIssuesCount: Scalars['Int']['output'];
  createdAt: Scalars['ISO8601DateTime']['output'];
  creator?: Maybe<Owner>;
  description?: Maybe<Scalars['String']['output']>;
  dueOn?: Maybe<Scalars['ISO8601DateTime']['output']>;
  ghCreatedAt: Scalars['ISO8601DateTime']['output'];
  ghId: Scalars['Int']['output'];
  ghNodeId: Scalars['ID']['output'];
  ghUpdatedAt: Scalars['ISO8601DateTime']['output'];
  id: Scalars['ID']['output'];
  issues: IssueConnection;
  number: Scalars['Int']['output'];
  openIssuesCount: Scalars['Int']['output'];
  repository: Repository;
  startDate?: Maybe<Scalars['ISO8601DateTime']['output']>;
  state: MilestoneState;
  title: Scalars['String']['output'];
  updatedAt: Scalars['ISO8601DateTime']['output'];
};
// ...
export type Node = {
  /** ID of the object. */
  id: Scalars['ID']['output'];
};
// ...
export type Owner = {
  avatarUrl: Scalars['String']['output'];
  createdAt: Scalars['ISO8601DateTime']['output'];
  ghId?: Maybe<Scalars['Int']['output']>;
  hasWorkspace: Scalars['Boolean']['output'];
  id: Scalars['ID']['output'];
  login: Scalars['String']['output'];
  projects: ProjectConnection;
  /** Repositories favorited by this github user */
  repositoryFavorites: RepositoryFavoriteConnection;
  roadmapItems: RoadmapItemConnection;
  searchWorkspaces: WorkspaceSearchMatchConnection;
  type: Scalars['String']['output'];
  updatedAt: Scalars['ISO8601DateTime']['output'];
  workspaceFavorites: WorkspaceFavoriteConnection;
};
// ...
export type Pipeline = Node & Timestamps & {
  __typename?: 'Pipeline';
  createdAt: Scalars['ISO8601DateTime']['output'];
  description?: Maybe<Scalars['String']['output']>;
  hasEstimatedIssues: Scalars['Boolean']['output'];
  id: Scalars['ID']['output'];
  /** Whether the pipeline is a default for new PRs */
  isDefaultPRPipeline: Scalars['Boolean']['output'];
  /** @deprecated This field must not be used. */
  isEpicPipeline: Scalars['Boolean']['output'];
  issues: IssueConnection;
  itemBefore?: Maybe<Pipeline>;
  name: Scalars['String']['output'];
  pipelineConfiguration: PipelineConfiguration;
  pipelineToPipelineAutomationDestinations: PipelineToPipelineAutomationConnection;
  pipelineToPipelineAutomationSources: PipelineToPipelineAutomationConnection;
  stage?: Maybe<PipelineStage>;
  updatedAt: Scalars['ISO8601DateTime']['output'];
  workspace: Workspace;
};
// ...
export type PipelineIssue = {
  __typename?: 'PipelineIssue';
  id: Scalars['ID']['output'];
  issue: Issue;
  itemAfter?: Maybe<PipelineIssue>;
  itemBefore?: Maybe<PipelineIssue>;
  latestTransferTime?: Maybe<Scalars['ISO8601DateTime']['output']>;
  pipeline: Pipeline;
  priority?: Maybe<Priority>;
  relativePosition: Scalars['BigInt']['output'];
  workspace: Workspace;
};
// ...
export type PipelineToPipelineAutomationConnection = {
  __typename?: 'PipelineToPipelineAutomationConnection';
  /** A list of edges. */
  edges: Array<PipelineToPipelineAutomationEdge>;
  /** A list of nodes. */
  nodes: Array<PipelineToPipelineAutomation>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type Query = {
  __typename?: 'Query';
  /** Lookup an Issue by repository_gh_id or repository_id  and issue_number */
  issueByInfo: Issue;
  /**
   * A concatenated list of all issues connected to any
   *            of the given issues in the passed workspace
   */
  issueConnections: IssueConnection;
  milestoneByRepoGhIdAndNumber: Milestone;
  /** Fetches an object given its ID. */
  node?: Maybe<Node>;
  /** Fetches a list of objects given a list of IDs. */
  nodes: Array<Maybe<Node>>;
  /** Lookup an Owner by GitHub database ID */
  ownerByGhId?: Maybe<Owner>;
  /** Lookup an Owner by GitHub login */
  ownerByLogin?: Maybe<Owner>;
  /** preview sprint config schedules */
  previewSprintConfig?: Maybe<SprintConfigPreview>;
  /** User's recently viewed Workspaces */
  recentlyViewedWorkspaces: WorkspaceConnection;
  /** Lookup repositories by their GitHub ID */
  repositoriesByGhId: Array<Repository>;
  /** Search and filter closed issues per pipeline */
  searchClosedIssues?: Maybe<IssueConnection>;
  /** Search and filter issues inside repository */
  searchIssues?: Maybe<IssueConnection>;
  /** Search and filter issues per pipeline */
  searchIssuesByPipeline?: Maybe<IssueConnection>;
  /** Search and filter issues inside zenhub epics */
  searchIssuesByZenhubEpics?: Maybe<IssueConnection>;
  /** PipelineToPipelineAutomation records that are triggered when an issue moves into a Pipeline */
  triggeredPipelineToPipelineAutomations: PipelineToPipelineAutomationConnection;
  /** Validate a workspace name within a zorg */
  validateWorkspaceName: Scalars['Boolean']['output'];
  /** Return the current zenhub user */
  viewer: ZenhubUser;
  /** Lookup Workspace by its ID */
  workspace?: Maybe<Workspace>;
};
// ...
export type Sprint = Node & Timestamps & {
  __typename?: 'Sprint';
  closedIssuesCount: Scalars['Int']['output'];
  completedPoints: Scalars['Float']['output'];
  createdAt: Scalars['ISO8601DateTime']['output'];
  description?: Maybe<Scalars['String']['output']>;
  endAt: Scalars['ISO8601DateTime']['output'];
  generatedName?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  issues: IssueConnection;
  name?: Maybe<Scalars['String']['output']>;
  persisted: Scalars['Boolean']['output'];
  /** batching is disabled for this field */
  scopeChange: ScopeChangeConnection;
  sprintIssues: SprintIssueConnection;
  sprintReview?: Maybe<SprintReview>;
  startAt: Scalars['ISO8601DateTime']['output'];
  state: SprintState;
  totalPoints: Scalars['Float']['output'];
  updatedAt: Scalars['ISO8601DateTime']['output'];
  workspace: Workspace;
};
// ...
export type SprintConfigPreview = {
  __typename?: 'SprintConfigPreview';
  periodInWeeks: Scalars['Int']['output'];
  sprints: Array<Sprint>;
};
// ...
export type SprintConnection = {
  __typename?: 'SprintConnection';
  /** A list of edges. */
  edges: Array<SprintEdge>;
  /** A list of nodes. */
  nodes: Array<Sprint>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type User = Owner & Timestamps & {
  __typename?: 'User';
  avatarUrl: Scalars['String']['output'];
  createdAt: Scalars['ISO8601DateTime']['output'];
  ghId?: Maybe<Scalars['Int']['output']>;
  hasWorkspace: Scalars['Boolean']['output'];
  id: Scalars['ID']['output'];
  login: Scalars['String']['output'];
  name?: Maybe<Scalars['String']['output']>;
  projects: ProjectConnection;
  /** Repositories favorited by this github user */
  repositoryFavorites: RepositoryFavoriteConnection;
  roadmapItems: RoadmapItemConnection;
  searchWorkspaces: WorkspaceSearchMatchConnection;
  type: Scalars['String']['output'];
  updatedAt: Scalars['ISO8601DateTime']['output'];
  workspaceFavorites: WorkspaceFavoriteConnection;
  zenhubUser?: Maybe<ZenhubUser>;
};
// ...
export type UserConnection = {
  __typename?: 'UserConnection';
  /** A list of edges. */
  edges: Array<UserEdge>;
  /** A list of nodes. */
  nodes: Array<User>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type WorkspaceConnection = {
  __typename?: 'WorkspaceConnection';
  /** A list of edges. */
  edges: Array<WorkspaceEdge>;
  /** A list of nodes. */
  nodes: Array<Workspace>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type ChrysalisGetEpicLinkedIssuesQuery = { __typename?: 'Query', linkedIssues?: { __typename?: 'IssueConnection', nodes: Array<{ __typename?: 'Issue', number: number, title: string, htmlUrl: string, state: IssueState, assignees: { __typename?: 'UserConnection', nodes: Array<{ __typename?: 'User', login: string }> }, blockingIssues: { __typename?: 'IssueConnection', nodes: Array<{ __typename?: 'Issue', number: number }> }, blockedIssues: { __typename?: 'IssueConnection', nodes: Array<{ __typename?: 'Issue', number: number }> }, pipelineIssue?: { __typename?: 'PipelineIssue', pipeline: { __typename?: 'Pipeline', name: string } } | null, estimate?: { __typename?: 'Estimate', value: number } | null, sprints: { __typename?: 'SprintConnection', nodes: Array<{ __typename?: 'Sprint', name?: string | null }> } }> } | null };